LRU8BIT and OPTIMAL scheduling added
USAGE - Include the header file LRU8BIT_OPTIMAL.h, then:
LRU_8BIT(int refList[], int refCount, in frameCount) and OPTIMAL (int
refList[], int refCount, int frameCount) are used with same args:
refList is the references read from the input file as an int array.
refCount is the amount of elements in refList.
frameCount is the amount of frames available as requested by command
line arg

hasRef(int frameArray[], int frameCount, int ref) is meant to be used
within this code only: given an int array, the size of said array,  and
an int, it either tells you 1 + the index of that int, or 0 if that int
is not in the array.

That's all there is to it.